Visual feedback working for provisioning and update
Visual feed back working for provisioning completely, the functionality is as follows;
When connecting to Wi-Fi the yellow light turns on, if it connects the green light turns on and yellow light turns off. If it fails to connect the red light turns on and the yellow light turns off.

The OTA update visual feedback is there but doesn't do exactly what I want so will do a bug fix in following commit.
Currently, yellow light turns on when updating  and green/ red LED blinks when updated or failure in similar way to provisioning (except with blinking this time) but the problem is the yellow light doesn't blink and doesn't turn off when it should.