User can now define shuffling speed with slider from interface.
Changes made:
Added a slider to the interface that allows the user to choose the speed the motors run at.
This was challenging as I needed to concatenate the number of cards and speed together as an output from the interface and then parse them afterwards to separate them as the command function in the aREST interface only allows one output at a time.