Fixed length of time motors run proportional to defined speed and number of cards
Changes made:
The user defined speed caused the motors to run at different speeds but did not take into account that this would affect the number of cards shuffled. The length of time the motors run is now proportional to the speed defined by the user so it runs for the correct amount of time for the correct amount of cards that the user defines.