Fixed Custom and Riffle shuffle functions on bootstrap interface.
Bootstrap interface caused only two out of 4 buttons to work which went unnoticed originally but was discovered during testing now. The text boxes had been set up as buttons and therefore the buttons related to those text boxes were not functional. This is now fixed and the whole interface works as intended.