Added Video Demonstration of Card Shuffler
Added a 50 second long video demonstration of the interface and card shuffler in action to demonstrate what it can do. This commit also adds a "Documentation Resources" folder for adding future assets for documentation such as images and the 3D print STL files. He video is 720p not HD because GitHub wouldn't let me push a file larger than 100MB